backend/requirements.txt
asgiref==3.9.2
dj-database-url==3.0.1
Django==5.2.7
django-allauth==65.11.2
django-cors-headers==4.9.0
djangorestframework==3.16.1
gunicorn==23.0.0
packaging==25.0
pillow==11.3.0
psycopg2-binary==2.9.10
PyJWT==2.10.1
sqlparse==0.5.3
whitenoise==6.11.0
